+++ test/example/paramAttribute/index.js
@@ -1,4 +1,4 @@
 module.exports = {
-    paramsAlt: 'id',
+    paramAlt: 'id',
     get: function(req, res, next) {} // jshint ignore:line
 };