+++ test/example/params/multiple.js
@@ -1,4 +1,4 @@
 module.exports = {
-    params: ['id1', 'id2'],
+    param: ['id1', 'id2'],
     get: function(req, res, next) {} // jshint ignore:line
 };