+++ test/example/pathPart/view.js
@@ -1,5 +1,5 @@
 module.exports = {
-    params: 'id',
+    param: 'id',
     pathPart: null,
     get: function(req, res, next) {} // jshint ignore:line
 };